+++ js/core.js
@@ -29,16 +29,6 @@ var Toolkit = {
     /** Does the browser support transitions? */
     hasTransition: hasTransition,
 
-    /** Event name for transition end */
-    transitionEnd: (function() {
-        var eventMap = {
-            WebkitTransition: 'webkitTransitionEnd',
-            OTransition: 'oTransitionEnd otransitionend'
-        };
-
-        return eventMap[hasTransition] || 'transitionend';
-    })(),
-
     /** Detect touch devices */
     isTouch: isTouch,
 
@@ -49,16 +39,16 @@ var Toolkit = {
     cache: {},
 
     /**
-     * Creates a jQuery plugin by extending the jQuery prototype and defines a method
-     * that initializes a component. The component is only initialized if one has not been already.
-     * Components are either defined per element, or on a collection of elements.
+     * Creates a jQuery plugin by extending the jQuery prototype with a method definition.
+     * The Toolkit plugin is only initialized if one has not been already.
+     * Plugins are either defined per element, or on a collection of elements.
      *
-     * @param {String} component
+     * @param {String} plugin
      * @param {Function} callback
      * @param {bool} collection
      */
-    create: function(component, callback, collection) {
-        var name = component;
+    create: function(plugin, callback, collection) {
+        var name = plugin;
 
         // Prefix with toolkit to avoid collisions
         if ($.fn[name]) {
@@ -69,10 +59,10 @@ var Toolkit = {
 
             // Apply the instance to a collection of elements
             function() {
-                var instance = Toolkit.cache[component + '.' + this.selector] = callback.apply(this, arguments);
+                var instance = Toolkit.cache[plugin + '.' + this.selector] = callback.apply(this, arguments);
 
                 return this.each(function() {
-                    $(this).cache('toolkit.' + component, instance);
+                    $(this).cache('toolkit.' + plugin, instance);
                 });
             } :
 
@@ -81,7 +71,7 @@ var Toolkit = {
                 var args = arguments;
 
                 return this.each(function() {
-                    $(this).cache('toolkit.' + component, callback.apply(this, args));
+                    $(this).cache('toolkit.' + plugin, callback.apply(this, args));
                 });
             };
     }